Selection de données en fonction de critère temporels et de structure Applicable sur des objets en sortie de pmeasyr

Selection de données en fonction de critère temporels et de structure Applicable sur des objets en sortie de pmeasyr

get_data(DF, ref = "ansor", m, a, val = NULL, niveau = NULL,
  opt = T)

Arguments

ref
chaine de caractères, la variable de référence qui sera utilisée pour la selection temporel - si ref = ansor : les variables pmeasyr ansor et moisor sont utilisé pour la sélection avec les paramètres m et a par exemple pour des objets de type rum ou rsa

m
vecteur numérique, un vecteur de nombre entiers indiquant les mois à selectionner

a
vecteur numérique, un vecteur de nombre entiers indiquant les années à selectionner

val
chaine de caractères, variable géographique qui sera choisie pour la selection (ex : service, pole,...). Si NULL pas de critères géographiques pour la sélection

niveau
chaine de caractères, valeur de la variable géographique à sélectionner

df
tibble, un tibble en sortie de pmeasyr

regles
logique, utilisation des règles locales T

Value

tibble, un tibble contenant la sélection selon les critères indiqués