Fonction pmctmono (on doit encore repartir les supplements RAP et ANTEPARTUM)

Détails du nom des variables pmeasyr en entrée (rsa valorisés) rec_totale :Valorisation 100 rec_base: Valorisation des GHS de base rec_bee: Valorisation base + exb + exh rec_exb: Valorisation extrême bas (à déduire) rec_rehosp_ghm: Valorisation séjours avec rehosp dans même GHM rec_mino_sus: Valorisation séjours avec minoration forfaitaire liste en sus rec_exh: Valorisation journées extrême haut rec_aph: Valorisation actes GHS 9615 en Hospit. rec_rap: Valorisation suppléments radiothérapie pédiatrique, rec_ant: Valorisation suppléments antepartum, rec_rdt_tot: Valorisation actes RDTH en Hospit., rec_rea: Valorisation suppléments de réanimation rec_rep: Valorisation suppléments de réa pédiatrique rec_nn1: Valorisation suppléments de néonat sans SI rec_nn2: Valorisation suppléments de néonat avec SI rec_nn3: Valorisation suppléments de réanimation néonat rec_po_tot: Valorisation prélévements d organe rec_caishyp: Valorisation des actes de caissons hyperbares en sus rec_dialhosp: Valorisation suppléments de dialyse rec_src: Valorisation suppléments de surveillance continue rec_stf: Valorisation suppléments de soins intensifs rec_sdc: Valorisation supplément défibrilateur cardiaque

pmct_mono_uma(rsa, rum, annee, mois)

Arguments

rsa
rsa valorisés = tibble vrsa$rsa en sortie de pmeasyr::vvr_ghs_supp

rum
rum non valorisés = tibble rum$rum en sortie de pmeasyr::irum

pmctmono
pmct mono uma = tibble en sortie dimRactivite::pmct_mono_uma

Value

rum_v = rum valorisés

Details

La répartition des recettes par RUM est réalisée par la fonction vvr_rum_repa. Plusieurs clés ont été proposées pour cette répartition. Le package en implémente 4 qui sont disponibles dans l'objet rum_v (les valeur et les coefficients sont disponibles) : - répartition selon la durée du RUM (coeftime,valotime) - répartition selon le pmct moyen dans l'unité pour les mono RUM (coefpmctmono,valopmctmono) - répartition selon un coefficient composite obtenu par multiplication des deux précédents (coefpmctmonotime1,valopmctmonotime1) - répartition selon un coefficient composite obtenu par la moyenne des deux précédents (coefpmctmonotime2,valopmctmonotime2) L'ensemble des suppléments individuels sont présents en sortie avec le suffixes _repa

Examples